raspberrypi/requirements.txt
opencv-python
pyserial
matplotlib
gpiozero
RPi.GPIO
lgpio